# src/core/Snakk.Sdk/openapi.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "Snakk.Api | v1",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "https://localhost:17100/"
    }
  ],
  "paths": {
...
    "/hubs/{hubId}/spaces": {
      "get": {
        "tags": [
          "Hubs"
        ],
        "operationId": "GetSpacesByHub",
        "parameters": [
          {
            "name": "hubId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": true,
            "schema": {
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": [
                "integer",
                "string"
              ],
              "format": "int32"
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "required": true,
            "schema": {
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": [
                "integer",
                "string"
              ],
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PagedResponseOfSpaceByHubResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/spaces": {
      "post": {
        "tags": [
          "Spaces"
        ],
        "operationId": "CreateSpace",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateSpaceRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SpaceResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/auth/update-profile": {
      "put": {
        "tags": [
          "Authentication"
        ],
        "operationId": "UpdateProfile",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProfileRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UpdateProfileResponse"
                }
              }
            }
          }
        }
      }
    },
    "/auth/preferences": {
      "put": {
        "tags": [
          "Authentication"
        ],
        "operationId": "UpdatePreferences",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePreferencesRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/c/{slug}/manage/overview": {
      "get": {
        "tags": [
          "Community Management"
        ],
        "operationId": "GetCommunityOverview",
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/c/{slug}/manage/settings": {
      "get": {
        "tags": [
          "Community Management"
        ],
        "operationId": "GetCommunitySettings",
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "put": {
        "tags": [
          "Community Management"
        ],
        "operationId": "UpdateCommunitySettings",
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCommunitySettingsRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/c/{slug}/manage/moderation": {
      "get": {
        "tags": [
          "Community Management"
        ],
        "operationId": "GetCommunityModeration",
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/c/{slug}/manage/spaces": {
      "get": {
        "tags": [
          "Community Management"
        ],
        "operationId": "GetCommunitySpaces",
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/c/{communitySlug}/h/{hubSlug}/manage/overview": {
      "get": {
        "tags": [
          "Hub Management"
        ],
        "operationId": "GetHubOverview",
        "parameters": [
          {
            "name": "communitySlug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "hubSlug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/c/{communitySlug}/h/{hubSlug}/manage/settings": {
      "get": {
        "tags": [
          "Hub Management"
        ],
        "operationId": "GetHubSettings",
        "parameters": [
          {
            "name": "communitySlug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "hubSlug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "put": {
        "tags": [
          "Hub Management"
        ],
        "operationId": "UpdateHubSettings",
        "parameters": [
          {
            "name": "communitySlug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "hubSlug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateHubSettingsRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/c/{communitySlug}/h/{hubSlug}/manage/moderation": {
      "get": {
        "tags": [
          "Hub Management"
        ],
        "operationId": "GetHubModeration",
        "parameters": [
          {
            "name": "communitySlug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "hubSlug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/c/{communitySlug}/h/{hubSlug}/manage/spaces": {
      "get": {
        "tags": [
          "Hub Management"
        ],
        "operationId": "GetHubSpaces",
        "parameters": [
          {
            "name": "communitySlug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "hubSlug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/c/{communitySlug}/s/{spaceSlug}/manage/overview": {
      "get": {
        "tags": [
          "Space Management"
        ],
        "operationId": "GetSpaceOverview",
        "parameters": [
          {
            "name": "communitySlug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "spaceSlug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/c/{communitySlug}/s/{spaceSlug}/manage/settings": {
      "get": {
        "tags": [
          "Space Management"
        ],
        "operationId": "GetSpaceSettings",
        "parameters": [
          {
            "name": "communitySlug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "spaceSlug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "put": {
        "tags": [
          "Space Management"
        ],
        "operationId": "UpdateSpaceSettings",
        "parameters": [
          {
            "name": "communitySlug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "spaceSlug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateSpaceSettingsRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/c/{communitySlug}/s/{spaceSlug}/manage/moderation": {
      "get": {
        "tags": [
          "Space Management"
        ],
        "operationId": "GetSpaceModeration",
        "parameters": [
          {
            "name": "communitySlug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "spaceSlug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/c/{communitySlug}/s/{spaceSlug}/manage/rules": {
      "get": {
        "tags": [
          "Space Management"
        ],
        "operationId": "GetSpaceRulesManagement",
        "parameters": [
          {
            "name": "communitySlug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "spaceSlug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "put": {
        "tags": [
          "Space Management"
        ],
        "operationId": "UpdateSpaceRules",
        "parameters": [
          {
            "name": "communitySlug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "spaceSlug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateSpaceRulesRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CreateSpaceRequest": {
        "required": [
          "hubId",
          "name",
          "slug",
          "description"
        ],
        "type": "object",
        "properties": {
          "hubId": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "slug": {
            "type": "string"
          },
          "description": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
...
      "LatestDiscussionRef": {
        "required": [
          "publicId",
          "title",
          "slug",
          "lastActivityAt",
          "authorPublicId",
          "authorDisplayName",
          "authorAvatarFileName",
          "postCount"
        ],
        "type": "object",
        "properties": {
          "publicId": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "slug": {
            "type": "string"
          },
          "lastActivityAt": {
            "type": "string",
            "format": "date-time"
          },
          "authorPublicId": {
            "type": "string"
          },
          "authorDisplayName": {
            "type": "string"
          },
          "authorAvatarFileName": {
            "type": [
              "null",
              "string"
            ]
          },
          "postCount": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          }
        }
      },
...
      "MessageResponse": {
        "required": [
          "message"
        ],
        "type": "object",
        "properties": {
          "message": {
            "type": "string"
          }
        }
      },
...
      "PagedResponseOfSpaceByHubResponse": {
        "required": [
          "items",
          "offset",
          "pageSize",
          "hasMoreItems"
        ],
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SpaceByHubResponse"
            }
          },
          "offset": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "pageSize": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "hasMoreItems": {
            "type": "boolean"
          },
          "nextCursor": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
...
      "SpaceByHubResponse": {
        "required": [
          "publicId",
          "hubPublicId",
          "name",
          "slug",
          "description",
          "createdAt",
          "discussionCount",
          "replyCount"
        ],
        "type": "object",
        "properties": {
          "publicId": {
            "type": "string"
          },
          "hubPublicId": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "slug": {
            "type": "string"
          },
          "description": {
            "type": [
              "null",
              "string"
            ]
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "discussionCount": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "replyCount": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "latestDiscussion": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/LatestDiscussionRef"
              }
            ]
          }
        }
      },
...
      "SpaceResponse": {
        "required": [
          "publicId",
          "hubId",
          "name",
          "slug",
          "description",
          "createdAt",
          "discussionCount",
          "replyCount"
        ],
        "type": "object",
        "properties": {
          "publicId": {
            "type": "string"
          },
          "hubId": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "slug": {
            "type": "string"
          },
          "description": {
            "type": [
              "null",
              "string"
            ]
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "discussionCount": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "replyCount": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "lastModifiedAt": {
            "type": [
              "null",
              "string"
            ],
            "format": "date-time"
          },
          "latestDiscussion": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/LatestDiscussionRef"
              }
            ]
          }
        }
      },
      "SpaceRuleDto": {
        "required": [
          "title",
          "description"
        ],
        "type": "object",
        "properties": {
          "id": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "title": {
            "maxLength": 100,
            "minLength": 0,
            "type": "string"
          },
          "description": {
            "maxLength": 500,
            "minLength": 0,
            "type": "string"
          },
          "order": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          }
        }
      },
...
      "UpdateCommunitySettingsRequest": {
        "required": [
          "name"
        ],
        "type": "object",
        "properties": {
          "name": {
            "maxLength": 100,
            "minLength": 3,
            "type": "string"
          },
          "description": {
            "maxLength": 500,
            "minLength": 0,
            "type": [
              "null",
              "string"
            ]
          },
          "avatarUrl": {
            "type": [
              "null",
              "string"
            ]
          },
          "customDomain": {
            "type": [
              "null",
              "string"
            ]
          },
          "isPrivate": {
            "type": "boolean"
          },
          "requireApproval": {
            "type": "boolean"
          },
          "allowMemberInvites": {
            "type": "boolean"
          }
        }
      },
      "UpdateHubSettingsRequest": {
        "required": [
          "name"
        ],
        "type": "object",
        "properties": {
          "name": {
            "maxLength": 100,
            "minLength": 3,
            "type": "string"
          },
          "description": {
            "maxLength": 500,
            "minLength": 0,
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
...
      "UpdatePreferencesRequest": {
        "required": [
          "preferEndlessScroll",
          "autoFollowOnReply"
        ],
        "type": "object",
        "properties": {
          "preferEndlessScroll": {
            "type": [
              "null",
              "boolean"
            ]
          },
          "autoFollowOnReply": {
            "type": [
              "null",
              "boolean"
            ]
          }
        }
      },
      "UpdateProfileRequest": {
        "required": [
          "displayName"
        ],
        "type": "object",
        "properties": {
          "displayName": {
            "type": "string"
          }
        }
      },
      "UpdateProfileResponse": {
        "required": [
          "message"
        ],
        "type": "object",
        "properties": {
          "message": {
            "type": "string"
          },
          "token": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
      "UpdateSpaceRulesRequest": {
        "type": "object",
        "properties": {
          "rules": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SpaceRuleDto"
            }
          }
        }
      },
      "UpdateSpaceSettingsRequest": {
        "required": [
          "name"
        ],
        "type": "object",
        "properties": {
          "name": {
            "maxLength": 100,
            "minLength": 3,
            "type": "string"
          },
          "description": {
            "maxLength": 500,
            "minLength": 0,
            "type": [
              "null",
              "string"
            ]
          },
          "allowedThreadTypes": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "requireApproval": {
            "type": "boolean"
          },
          "allowAnonymous": {
            "type": "boolean"
          }
        }
      },
...
    }
  },
...
}